requirements.txt
flask
flask_pymongo
flask_cors
flask_jwt_extended
flask_bcrypt
cloudinary
Pillow
tensorflow
matplotlib
